Fix class rename
