+++ services/worker/src/worker/job_runners/_cached_directory_job_runner.py
